- translated nintendo platforms - _characterProvider updates when amiibo changes (ignoring wished and owned) - Game Header change color based on pinned
