+++ .gitignore
@@ -97,3 +97,7 @@ lib/**/*.freezed.dart
 # **/android/app/main/res/**/*background.png
 
 local.properties
+
+# Files generated by Flutter Intl
+lib/generated/**/**
+test/**/*.mocks.dart
