+++ lib/widget/game_list.dart
@@ -1,3 +1,4 @@
+import 'package:amiibo_network/generated/l10n.dart';
 import 'package:amiibo_network/model/game.dart';
 import 'package:amiibo_network/riverpod/amiibo_provider.dart';
 import 'package:amiibo_network/riverpod/game_provider.dart';
@@ -13,6 +14,7 @@ class GameListWidget extends ConsumerWidget {
   @override
   Widget build(BuildContext context, ScopedReader watch) {
     final id = watch(keyAmiiboProvider);
+    final S translate = S.of(context);
     return watch(gameProvider(id)).when(
       data: (platforms) {
         return MultiSliver(
@@ -20,20 +22,20 @@ class GameListWidget extends ConsumerWidget {
             if (platforms.games3DS != null && platforms.games3DS!.isNotEmpty)
               _PlatformGameList(
                 games: platforms.games3DS!,
-                title: '3DS',
+                title: translate.console_3DS_platform,
                 asset: 'ds',
               ),
             if (platforms.gamesWiiU != null && platforms.gamesWiiU!.isNotEmpty)
               _PlatformGameList(
                 games: platforms.gamesWiiU!,
-                title: 'WiiU',
+                title: translate.wiiu_platform,
                 asset: 'wiiu',
               ),
             if (platforms.gamesSwitch != null &&
                 platforms.gamesSwitch!.isNotEmpty)
               _PlatformGameList(
                 games: platforms.gamesSwitch!,
-                title: 'Switch',
+                title: translate.switch_platform,
                 asset: 'switch',
               ),
           ],
@@ -42,8 +44,8 @@ class GameListWidget extends ConsumerWidget {
       loading: () => const SliverToBoxAdapter(child: LinearProgressIndicator()),
       error: (e, _) {
         late final Widget child;
-        if (e is DioError && e.type == DioErrorType.response) {
-          if (e.response != null)
+        if (e is DioError) {
+          if (e.type == DioErrorType.response && e.response != null)
             switch (e.response!.statusCode) {
               case 404:
                 child = Text('No games for this amiibo yet');
@@ -56,7 +58,8 @@ class GameListWidget extends ConsumerWidget {
             child = Text(e.message);
         } else if (e is ArgumentError) {
           child = Text('Invalid amiibo data');
-        }
+        } else
+          child = Text(e.toString());
         return SliverToBoxAdapter(
           child: Center(
             child: DefaultTextStyle.merge(
@@ -95,6 +98,7 @@ class _PlatformGameList extends StatelessWidget {
           subtitle = Text(
             games[index].usage!.first.uses,
             style: theme.primaryTextTheme.subtitle2,
+            maxLines: 2,
           );
         }
         return DecoratedBox(
@@ -124,23 +128,8 @@ class _PlatformGameList extends StatelessWidget {
         SliverPersistentHeader(
           pinned: true,
           delegate: _SliverAnimatedPersistentHeader(
-            child: ListTile(
-              leading: Image.asset(
-                'assets/images/$asset.png',
-                cacheHeight: 24,
-                cacheWidth: 24,
-                height: 24,
-                width: 24,
-                colorBlendMode: BlendMode.srcATop,
-                color: theme.primaryColorBrightness == Brightness.dark
-                    ? Colors.white70
-                    : null,
-              ),
-              title: Text(
-                title,
-                style: theme.textTheme.headline6,
-              ),
-            ),
+            title: title,
+            asset: asset,
           ),
         ),
         SliverPadding(
@@ -157,11 +146,13 @@ class _PlatformGameList extends StatelessWidget {
 class _SliverAnimatedPersistentHeader extends SliverPersistentHeaderDelegate {
   final double _maxExtent;
   final double _minExtent;
-  final Widget child;
+  final String title;
+  final String asset;
 
   const _SliverAnimatedPersistentHeader({
     this.vsync,
-    required this.child,
+    required this.title,
+    required this.asset,
   })  : _maxExtent = kToolbarHeight,
         _minExtent = kToolbarHeight,
         super();
@@ -170,16 +161,41 @@ class _SliverAnimatedPersistentHeader extends SliverPersistentHeaderDelegate {
   Widget build(
       BuildContext context, double shrinkOffset, bool overlapsContent) {
     final theme = Theme.of(context);
+    final appBarTheme = AppBarTheme.of(context);
     final bool isPinned = shrinkOffset > maxExtent - minExtent;
     return SizedBox(
       height: math.max(minExtent, maxExtent - shrinkOffset),
       child: AnimatedPhysicalModel(
         duration: kThemeAnimationDuration,
-        child: child,
+        child: ListTileTheme.merge(
+          textColor:
+              isPinned ? appBarTheme.iconTheme?.color : theme.iconTheme.color,
+          iconColor: isPinned
+              ? appBarTheme.color ?? theme.iconTheme.color
+              : theme.iconTheme.color,
+          child: ListTile(
+            leading: Image.asset(
+              'assets/images/$asset.png',
+              cacheHeight: 24,
+              cacheWidth: 24,
+              height: 24,
+              width: 24,
+              colorBlendMode: BlendMode.srcATop,
+              color: isPinned
+                  ? appBarTheme.iconTheme?.color
+                  : theme.iconTheme.color,
+            ),
+            title: Text(
+              title,
+            ),
+          ),
+        ),
         shape: BoxShape.rectangle,
         animateShadowColor: false,
         elevation: isPinned ? 8.0 : 0.0,
-        color: isPinned ? AppBarTheme.of(context).backgroundColor ?? theme.primaryColor : theme.cardColor,
+        color: isPinned
+            ? AppBarTheme.of(context).backgroundColor ?? theme.primaryColor
+            : theme.cardColor,
         shadowColor: theme.shadowColor,
       ),
     );
@@ -196,6 +212,6 @@ class _SliverAnimatedPersistentHeader extends SliverPersistentHeaderDelegate {
 
   @override
   bool shouldRebuild(_SliverAnimatedPersistentHeader oldDelegate) {
-    return child != oldDelegate.child;
+    return asset != oldDelegate.asset && title != oldDelegate.title;
   }
 }
